+++ src/services/auth.ts
@@ -1,59 +1,113 @@
-/* const API_BASE_URL = 'https://mobile.werate.io/api';
+import axios from 'axios';
+
+//authentication leaving commented code here to use it with a proxy
+/* 
 const proxyUrl = 'https://cors-anywhere.herokuapp.com/';
-
-async function register(username: string, password: string) {
-  const response = await fetch(proxyUrl + `${API_BASE_URL}/v1/register`, {
-    method: 'POST',
-    headers: { ['Content-Type']: 'application/json' },
-    body: JSON.stringify({ username, password })
-  });
-  return response.json();
-}
-
-async function login(username: string, password: string) {
-  const response = await fetch(proxyUrl + `${API_BASE_URL}/v2/login`, {
-    method: 'POST',
-    headers: { ['Content-Type']: 'application/json' },
-    body: JSON.stringify({ username, password })
-  });
-  const data = await response.json();
-
-  // Store tokens in cookies with expiration based on response, keeping this on hold to take pull changes
-  if (data.accessToken && data.refreshToken) {
-    const cookies = new Cookies(req, res);
-
-    // Calculate maxAge based on expiration dates
-    const accessExpirationDate = new Date(data.accessExpirationDate).getTime() - Date.now();
-    const refreshExpirationDate = new Date(data.refreshExpirationDate).getTime() - Date.now();
-
-    cookies.set('accessToken', data.accessToken, {
-      httpOnly: true,
-      secure: true,
-      sameSite: 'strict',
-      maxAge: accessExpirationDate
-    });
-    cookies.set('refreshToken', data.refreshToken, {
-      httpOnly: true,
-      secure: true,
-      sameSite: 'strict',
-      maxAge: refreshExpirationDate
-    });
+export const register = async (email: string, password: string) => {
+  try {
+    const response = await axios.post(
+      proxyUrl + `${API_BASE_URL}/v1/register`,
+      { email, password },
+      { headers: { ['Content-Type']: ['application/json'] } }
+    );
+    return response.data;
+  } catch (error) {
+    console.error('Error during registration:', error);
   }
+};
 
-  return data;
-}
+// Login function
+export const login = async (email: string, password: string) => {
+  try {
+    const response = await axios.post(
+      proxyUrl + `${API_BASE_URL}/v2/login`,
+      { email, password },
+      { headers: { ['Content-Type']: ['application/json'] } }
+    );
 
-async function checkMfa(preAuthToken: string, mfaCode: string) {
-  const response = await fetch(proxyUrl + `${API_BASE_URL}/v1/check-mfa`, {
-    method: 'POST',
-    headers: {
-      ['Content-Type']: 'application/json',
-      ['Authorization']: `Bearer ${preAuthToken}`
-    },
-    body: JSON.stringify({ code: mfaCode })
-  });
-  return response.json();
-}
+    const data = response.data;
 
-export { register, login, checkMfa };
+    // Assuming you want to store the token in localStorage
+    if (data.accessToken) {
+      localStorage.setItem('token', data.accessToken);
+    }
+
+    return data;
+  } catch (error) {
+    console.error('Error during login:', error);
+  }
+};
+
+// Check MFA function
+export const checkMfa = async (preAuthToken: string, mfaCode: string) => {
+  try {
+    const response = await axios.post(
+      proxyUrl + `${API_BASE_URL}/v1/check-mfa`,
+      { code: mfaCode },
+      {
+        headers: {
+          ['Content-Type']: ['application/json'],
+          ['Authorization']: `Bearer ${preAuthToken}`
+        }
+      }
+    );
+    return response.data;
+  } catch (error) {
+    console.error('Error during MFA check:', error);
+  }
+};
  */
+const API_BASE_URL = 'https://api.werate.io/api';
+export const register = async (email: string, password: string) => {
+  try {
+    const response = await axios.post(
+      `${API_BASE_URL}/v1/register`,
+      { email, password },
+      { headers: { ['Content-Type']: ['application/json'] } }
+    );
+    return response.data;
+  } catch (error) {
+    console.error('Error during registration:', error);
+  }
+};
+
+// Login function
+export const login = async (email: string, password: string) => {
+  try {
+    const response = await axios.post(
+      `${API_BASE_URL}/v2/login`,
+      { email, password },
+      { headers: { ['Content-Type']: ['application/json'] } }
+    );
+
+    const data = response.data;
+
+    // Assuming you want to store the token in localStorage
+    if (data.accessToken) {
+      localStorage.setItem('token', data.accessToken);
+    }
+
+    return data;
+  } catch (error) {
+    console.error('Error during login:', error);
+  }
+};
+
+// Check MFA function
+export const checkMfa = async (preAuthToken: string, mfaCode: string) => {
+  try {
+    const response = await axios.post(
+      `${API_BASE_URL}/v1/check-mfa`,
+      { code: mfaCode },
+      {
+        headers: {
+          ['Content-Type']: ['application/json'],
+          ['Authorization']: `Bearer ${preAuthToken}`
+        }
+      }
+    );
+    return response.data;
+  } catch (error) {
+    console.error('Error during MFA check:', error);
+  }
+};
